Add Caliente functionality
Add Caliente functionality is now working. It was creating a Json object
to be inserted in the database and setting the id as ZERO and causing an
duplicate key exception. Now it does not set the key as all, pk is
assigned by mongo automatically. Method was split in two, one to create
a new entry and another to update an existent entry.